Move SatisfiesConstraints
